Creating new src folder for modules; Adding all src files to folder
